feat: add reach hook
